refactor: CSPRNG, segment datatype, imports

+++ include/segment.hpp
@@ -14,21 +14,21 @@ struct Segment
     uint32_t acknowledgementNumber : 32;
     struct
     {
-        unsigned int data_offset : 4;
-        unsigned int reserved : 4;
+        uint8_t data_offset : 4;
+        uint8_t reserved : 4;
     };
 
     struct
     {
-        unsigned int cwr : 1;
+        uint8_t cwr : 1;
         // todo continue ...
-        unsigned int ece : 1;
-        unsigned int urg : 1;
-        unsigned int ack : 1;
-        unsigned int psh : 1;
-        unsigned int rst : 1;
-        unsigned int syn : 1;
-        unsigned int fin : 1;
+        uint8_t ece : 1;
+        uint8_t urg : 1;
+        uint8_t ack : 1;
+        uint8_t psh : 1;
+        uint8_t rst : 1;
+        uint8_t syn : 1;
+        uint8_t fin : 1;
     } flags;
 
     uint16_t window : 16;
@@ -72,7 +72,9 @@ Segment fin();
  */
 Segment finAck();
 
-// update return type as needed
+/**
+ * Calculate checksum of a segment
+ */
 uint8_t *calculateChecksum(Segment segment);
 
 /**
